% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesPLCox-package.R
\docType{package}
\name{BayesPLCox-package}
\alias{BayesPLCox}
\alias{BayesPLCox-package}
\title{BayesPLCox: Bayesian Cox Models via PL and GPL Likelihoods}
\description{
Bayesian inference for Cox regression models based on
Plackett-Luce (PL) and generalized/geometric Plackett-Luce (GPL)
likelihood formulations.
}
\author{
\strong{Maintainer}: Tomohiro Ohigashi \email{ohigashi@rs.tus.ac.jp}

}
\keyword{internal}
