% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_gplcox.R
\name{coef.gplcox}
\alias{coef.gplcox}
\title{Extract posterior mean coefficients from a GPL-Cox model}
\usage{
\method{coef}{gplcox}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"gplcox"}.}

\item{...}{Further arguments, currently ignored.}
}
\value{
A named numeric vector of posterior mean coefficients.
}
\description{
Extract posterior mean coefficients from a GPL-Cox model
}
