% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_plcox.R
\name{coef.plcox}
\alias{coef.plcox}
\title{Extract posterior mean coefficients from a PL-Cox model}
\usage{
\method{coef}{plcox}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"plcox"}.}

\item{...}{Further arguments, currently ignored.}
}
\value{
A named numeric vector of posterior mean coefficients.
}
\description{
Extract posterior mean coefficients from a PL-Cox model
}
