% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_gplcox.R
\name{confint.gplcox}
\alias{confint.gplcox}
\title{Credible intervals for a GPL-Cox model}
\usage{
\method{confint}{gplcox}(object, parm, level = 0.95, ...)
}
\arguments{
\item{object}{An object of class \code{"gplcox"}.}

\item{parm}{Parameter indices or names. Currently ignored.}

\item{level}{Credible interval level.}

\item{...}{Further arguments, currently ignored.}
}
\value{
A matrix of posterior credible intervals.
}
\description{
Computes equal-tailed posterior credible intervals for regression coefficients.
}
