% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_plcox.R
\name{confint.plcox}
\alias{confint.plcox}
\title{Extract posterior mean coefficients from a PL-Cox model}
\usage{
\method{confint}{plcox}(object, parm, level = 0.95, ...)
}
\arguments{
\item{object}{An object of class \code{"plcox"}.}

\item{parm}{Parameter indices or names. Currently ignored.}

\item{level}{Credible interval level.}

\item{...}{Further arguments, currently ignored.}
}
\value{
A named numeric vector of posterior mean coefficients.
}
\description{
Extract posterior mean coefficients from a PL-Cox model
}
