% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_diagnostics.R
\name{mcmc_diagnostics.bayescoxrank}
\alias{mcmc_diagnostics.bayescoxrank}
\title{MCMC Diagnostics for BayesPLCox Models}
\usage{
\method{mcmc_diagnostics}{bayescoxrank}(object, include_frailty = FALSE, ...)
}
\arguments{
\item{object}{A fitted object inheriting from class
\code{"bayescoxrank"}.}

\item{include_frailty}{Logical; whether diagnostics for individual
frailty effects should be included.}

\item{...}{Further arguments.}
}
\value{
A list containing MCMC diagnostics for regression
coefficients and, when applicable, frailty components.
}
\description{
MCMC Diagnostics for BayesPLCox Models
}
