% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_diagnostics.R
\name{mcmc_diagnostics}
\alias{mcmc_diagnostics}
\title{MCMC Diagnostics}
\usage{
mcmc_diagnostics(object, include_frailty = FALSE, ...)
}
\arguments{
\item{object}{A fitted model object.}

\item{include_frailty}{Logical; whether individual frailty effects
should also be included.}

\item{...}{Further arguments.}
}
\value{
A list of MCMC diagnostic tables.
}
\description{
Computes convergence and sampling-efficiency diagnostics for a
fitted BayesPLCox model.
}
