% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_general.R
\name{prepare_chain_inits}
\alias{prepare_chain_inits}
\title{Prepare Initial Values for Multiple Chains}
\usage{
prepare_chain_inits(init, chains)
}
\arguments{
\item{init}{Initial values supplied by the user.}

\item{chains}{Number of MCMC chains.}
}
\value{
A list of length \code{chains}.
}
\description{
Expands user-supplied initial values into one initialization
specification for each MCMC chain.
}
\keyword{internal}
