% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_gplcox.R
\name{print.gplcox}
\alias{print.gplcox}
\title{Print a fitted GPL-Cox model}
\usage{
\method{print}{gplcox}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"gplcox"}.}

\item{...}{Further arguments, currently ignored.}
}
\value{
The fitted model object, invisibly.
}
\description{
Prints a concise summary of a fitted \code{"gplcox"} object.
}
