% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_plcox.R
\name{print.plcox}
\alias{print.plcox}
\title{Print a fitted PL-Cox model}
\usage{
\method{print}{plcox}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"plcox"}.}

\item{...}{Further arguments, currently ignored.}
}
\value{
The fitted model object, invisibly.
}
\description{
Prints a concise summary of a fitted \code{"plcox"} object.
}
