% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppcheck_gplcox.R
\name{print.pp_check_tie_gplcox}
\alias{print.pp_check_tie_gplcox}
\title{Print a posterior predictive tie-block check}
\usage{
\method{print}{pp_check_tie_gplcox}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"pp_check_tie_gplcox"}.}

\item{...}{Further arguments, currently ignored.}
}
\value{
The object, invisibly.
}
\description{
Print a posterior predictive tie-block check
}
