% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_gplcox.R
\name{print.summary.gplcox}
\alias{print.summary.gplcox}
\title{Print a summary of a GPL-Cox model}
\usage{
\method{print}{summary.gplcox}(x, digits = 3, ...)
}
\arguments{
\item{x}{An object of class \code{"summary.gplcox"}.}

\item{digits}{Number of digits to print.}

\item{...}{Further arguments, currently ignored.}
}
\value{
The summary object, invisibly.
}
\description{
Print a summary of a GPL-Cox model
}
