% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_plcox.R
\name{print.summary.plcox}
\alias{print.summary.plcox}
\title{Print a summary of a PL-Cox model}
\usage{
\method{print}{summary.plcox}(x, digits = 3, ...)
}
\arguments{
\item{x}{An object of class \code{"summary.plcox"}.}

\item{digits}{Number of digits to print.}

\item{...}{Further arguments, currently ignored.}
}
\value{
The summary object, invisibly.
}
\description{
Print a summary of a PL-Cox model
}
