% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_surv.R
\name{sample_Z_geom}
\alias{sample_Z_geom}
\title{Sample latent geometric variables for GPL-Cox}
\usage{
sample_Z_geom(theta, R_sets)
}
\arguments{
\item{theta}{Vector of success probabilities.}

\item{R_sets}{A list of risk sets.}
}
\value{
An integer vector of latent geometric variables.
}
\description{
Internal helper function that samples latent geometric variables associated
with each risk set under the GPL-Cox model.
}
\keyword{internal}
