% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_gplcox.R
\name{summary.gplcox}
\alias{summary.gplcox}
\title{Summarize a fitted GPL-Cox model}
\usage{
\method{summary}{gplcox}(object, level = 0.95, ...)
}
\arguments{
\item{object}{An object of class \code{"gplcox"}.}

\item{level}{Credible interval level.}

\item{...}{Further arguments, currently ignored.}
}
\value{
An object of class \code{"summary.gplcox"}.
}
\description{
Computes posterior summaries for regression coefficients from a fitted
\code{"gplcox"} object.
}
