% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_plcox.R
\name{summary.plcox}
\alias{summary.plcox}
\title{Summarize a fitted PL-Cox model}
\usage{
\method{summary}{plcox}(object, level = 0.95, ...)
}
\arguments{
\item{object}{An object of class \code{"plcox"}.}

\item{level}{Credible interval level.}

\item{...}{Further arguments, currently ignored.}
}
\value{
An object of class \code{"summary.plcox"}.
}
\description{
Computes posterior summaries for regression coefficients from a fitted
\code{"plcox"} object.
}
