% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{traceplot}
\alias{traceplot}
\title{Trace plots for posterior samples}
\usage{
traceplot(x, ...)
}
\arguments{
\item{x}{A fitted model object.}

\item{...}{Further arguments passed to methods.}
}
\value{
A plot object, typically a \code{ggplot2} or \code{patchwork} object.
}
\description{
Generic function for trace plots of posterior samples from fitted Bayesian
Cox models.
}
