% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplcox_fit.R
\name{gplcox}
\alias{gplcox}
\title{Fit a Bayesian Cox model via the GPL-Cox likelihood}
\usage{
gplcox(
  formula,
  data,
  n_iter = 4000,
  burn = 2000,
  thin = 1,
  prior_mean = NULL,
  prior_var = 100,
  a_sig = 0.01,
  b_sig = 0.01,
  standardize = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{formula}{A formula with a \code{Surv()} response. A frailty term of the form
\code{(1 | group)} may be included.}

\item{data}{A data frame.}

\item{n_iter}{Total number of MCMC iterations.}

\item{burn}{Number of burn-in iterations.}

\item{thin}{Thinning interval.}

\item{prior_mean}{Prior mean vector for regression coefficients.}

\item{prior_var}{Prior variance for regression coefficients.}

\item{a_sig, b_sig}{Hyperparameters for the frailty variance prior.}

\item{standardize}{Logical; standardize covariates except the intercept.}

\item{verbose}{Logical; print progress messages?}
}
\value{
An object of class \code{"gplcox"}.
}
\description{
Fits a Bayesian Cox regression model based on the generalized Plackett-Luce
(GPL) likelihood using Gibbs sampling. The GPL-Cox model provides a generative
representation for tied event sets and can therefore be used for posterior
predictive checks of tie structures.
}
\examples{
\dontrun{
fit <- gplcox(Surv(time, status) ~ x1 + x2, data = dat)
summary(fit)
plot(fit)
traceplot(fit)
pp <- pp_check_tie.gplcox(fit)
plot(pp)
}

}
