% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_gplcox.R
\name{traceplot.gplcox}
\alias{traceplot.gplcox}
\title{Trace plots for a fitted GPL-Cox model}
\usage{
\method{traceplot}{gplcox}(x, intercept = FALSE, ...)
}
\arguments{
\item{x}{An object of class \code{"gplcox"}.}

\item{intercept}{Logical; include the intercept in the trace plot?}

\item{...}{Further arguments, currently ignored.}
}
\value{
A plot object.
}
\description{
Produces trace plots of posterior draws for regression coefficients and, when
present, frailty variance parameters.
}
