% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_plcox.R
\name{traceplot.plcox}
\alias{traceplot.plcox}
\title{Trace plots for a fitted PL-Cox model}
\usage{
\method{traceplot}{plcox}(x, intercept = FALSE, ...)
}
\arguments{
\item{x}{An object of class \code{"plcox"}.}

\item{intercept}{Logical; include the intercept in the trace plot?}

\item{...}{Further arguments, currently ignored.}
}
\value{
A plot object.
}
\description{
Produces trace plots of posterior draws for regression coefficients and, when
present, frailty variance parameters.
}
